+++ CHANGELOG.md
@@ -3,6 +3,72 @@
 All notable changes to this project will be documented in this file. See
 [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.10.5](https://github.com/carbon-design-system/carbon-charts/compare/v1.10.4...v1.10.5) (2023-06-27)
+
+### Bug Fixes
+
+- **1592:** cp styles from dist to packages/core
+  ([#1594](https://github.com/carbon-design-system/carbon-charts/issues/1594))
+  ([e6f1b98](https://github.com/carbon-design-system/carbon-charts/commit/e6f1b984d63f1dd6e040510687c4e510258e2a99))
+
+# Change Log
+
+All notable changes to this project will be documented in this file. See
+[Conventional Commits](https://conventionalcommits.org) for commit guidelines.
+
+## [1.10.4](https://github.com/carbon-design-system/carbon-charts/compare/v1.10.3...v1.10.4) (2023-06-27)
+
+**Note:** Version bump only for package @carbon/charts-monorepo
+
+# Change Log
+
+All notable changes to this project will be documented in this file. See
+[Conventional Commits](https://conventionalcommits.org) for commit guidelines.
+
+## [1.10.3](https://github.com/carbon-design-system/carbon-charts/compare/v1.10.2...v1.10.3) (2023-06-26)
+
+### Bug Fixes
+
+- **workflows:** downgrade versions again
+  ([#1589](https://github.com/carbon-design-system/carbon-charts/issues/1589))
+  ([72e229b](https://github.com/carbon-design-system/carbon-charts/commit/72e229bdbf718f351126f3ae17a3c6fc6f8befe2))
+
+# Change Log
+
+All notable changes to this project will be documented in this file. See
+[Conventional Commits](https://conventionalcommits.org) for commit guidelines.
+
+## [1.10.2](https://github.com/carbon-design-system/carbon-charts/compare/v1.10.1...v1.10.2) (2023-06-26)
+
+**Note:** Version bump only for package @carbon/charts-monorepo
+
+# Change Log
+
+All notable changes to this project will be documented in this file. See
+[Conventional Commits](https://conventionalcommits.org) for commit guidelines.
+
+## [1.10.1](https://github.com/carbon-design-system/carbon-charts/compare/v1.10.0...v1.10.1) (2023-06-26)
+
+### Bug Fixes
+
+- **package.json:** missing devDependencies for storybook
+  ([#1580](https://github.com/carbon-design-system/carbon-charts/issues/1580))
+  ([e208fcc](https://github.com/carbon-design-system/carbon-charts/commit/e208fcc77f2adee30800356f974e7f6d0a5103e1))
+
+# Change Log
+
+All notable changes to this project will be documented in this file. See
+[Conventional Commits](https://conventionalcommits.org) for commit guidelines.
+
+# [1.10.0](https://github.com/carbon-design-system/carbon-charts/compare/v1.9.0...v1.10.0) (2023-06-26)
+
+**Note:** Version bump only for package @carbon/charts-monorepo
+
+# Change Log
+
+All notable changes to this project will be documented in this file. See
+[Conventional Commits](https://conventionalcommits.org) for commit guidelines.
+
 # [1.9.0](https://github.com/carbon-design-system/carbon-charts/compare/v1.8.0...v1.9.0) (2023-06-26)
 
 **Note:** Version bump only for package @carbon/charts-monorepo